:recycle: convert time marker to svg
