--- src/components/VerticalMarker.css
@@ -1,21 +1,0 @@
-.vertical-marker_outer {
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  display: flex;
-  align-items: flex-end;
-}
-
-.vertical-marker_outer > * {
-  height: 340px;
-}
-
-.vertical-marker_left {
-  border-right: 1px solid tomato;
-}
-
-.vertical-marker_right {
-  flex-grow: 1;
-}
